requirements.txt
gspread
requests
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib 